Rename case directory
